move Contacts, Targets, AnghammaradException and refactor imports
