+++ anghammarad/src/main/scala/com/gu/anghammarad/messages/HangoutsService.scala
@@ -1,6 +1,6 @@
 package com.gu.anghammarad.messages
 
-import com.gu.anghammarad.AnghammaradException.Fail
+import com.gu.anghammarad.common.AnghammaradException.Fail
 import com.gu.anghammarad.models.HangoutMessage
 import sttp.client3.{HttpURLConnectionBackend, Response, UriContext, basicRequest}
