+++ anghammarad/src/main/scala/com/gu/anghammarad/messages/SendMessages.scala
@@ -1,7 +1,7 @@
 package com.gu.anghammarad.messages
 
-import com.gu.anghammarad.AnghammaradException.Fail
 import com.gu.anghammarad.Enrichments._
+import com.gu.anghammarad.common.AnghammaradException.Fail
 import com.gu.anghammarad.models._
 
 import scala.util.Try
