+++ anghammarad/src/main/scala/com/gu/anghammarad/serialization/Serialization.scala
@@ -2,7 +2,7 @@ package com.gu.anghammarad.serialization
 
 import cats.syntax.either._
 import com.amazonaws.services.lambda.runtime.events.SNSEvent
-import com.gu.anghammarad.AnghammaradException.Fail
+import com.gu.anghammarad.common.AnghammaradException.Fail
 import com.gu.anghammarad.Enrichments._
 import com.gu.anghammarad.models._
 import io.circe._
