+++ anghammarad/src/test/scala/com/gu/anghammarad/ContactsTest.scala
@@ -1,6 +1,6 @@
 package com.gu.anghammarad
 
-import com.gu.anghammarad.Contacts._
+import com.gu.anghammarad.common.Contacts._
 import com.gu.anghammarad.models._
 import com.gu.anghammarad.testutils.TryValues
 import org.scalatest.freespec.AnyFreeSpec
